Removes code taken from example form
