--- libs/application/templates/children-domicile-transfer/jest.config.js
@@ -1,13 +1,0 @@
-module.exports = {
-  name: 'application-template-reference-template',
-  preset: '../../../../jest.config.js',
-  transform: {
-    '^.+\\.[tj]sx?$': [
-      'babel-jest',
-      { cwd: __dirname, configFile: './babel-jest.config.json' },
-    ],
-  },
-  moduleFileExtensions: ['ts', 'tsx', 'js', 'jsx'],
-  coverageDirectory:
-    '../../../../coverage/libs/application/templates/reference-template',
-}
