--- libs/application/templates/children-domicile-transfer/src/fields/index.ts
@@ -1,2 +1,0 @@
-export { default as ExampleCountryField } from './ExampleCountryField'
-export { default as CustomRepeater } from './CustomRepeater'
