--- libs/application/templates/children-domicile-transfer/src/forms/Approved.ts
@@ -1,19 +1,0 @@
-import {
-  buildForm,
-  buildDescriptionField,
-  Form,
-  FormModes,
-} from '@island.is/application/core'
-
-export const Approved: Form = buildForm({
-  id: 'ApprovedApplicationForm',
-  title: 'Samþykkt',
-  mode: FormModes.APPROVED,
-  children: [
-    buildDescriptionField({
-      id: 'approved',
-      title: 'Til hamingju!',
-      description: 'Umsókn þín hefur verið samþykkt!',
-    }),
-  ],
-})
