--- libs/application/templates/children-domicile-transfer/src/forms/PendingReview.ts
@@ -1,19 +1,0 @@
-import {
-  buildForm,
-  buildDescriptionField,
-  Form,
-  FormModes,
-} from '@island.is/application/core'
-
-export const PendingReview: Form = buildForm({
-  id: 'ExamplePending',
-  title: 'Í vinnslu',
-  mode: FormModes.PENDING,
-  children: [
-    buildDescriptionField({
-      id: 'inReview',
-      title: 'Í vinnslu',
-      description: 'Umsókn þín um ökunám er nú í vinnslu. ',
-    }),
-  ],
-})
