--- libs/application/templates/children-domicile-transfer/src/forms/messages.ts
@@ -1,95 +1,0 @@
-import { defineMessages } from 'react-intl'
-
-export const m = defineMessages({
-  introSection: {
-    id: 'example.application:intro.section',
-    defaultMessage: 'Upplýsingar',
-    description: 'Some description',
-  },
-  introField: {
-    id: 'example.application:intro.field',
-    defaultMessage: 'Velkomin(n)',
-    description: 'Some description',
-  },
-  introIntroduction: {
-    id: 'example.application:intro.introduction',
-    defaultMessage:
-      '*Hello*, **{name}**! [This is a link to Google!](http://google.com)',
-    description: 'Some description',
-  },
-  about: {
-    id: 'example.application:about',
-    defaultMessage: 'Um þig',
-    description: 'Some description',
-  },
-  name: {
-    id: 'example.application:person.name',
-    defaultMessage: 'Nafn',
-    description: 'Some description',
-  },
-  nationalId: {
-    id: 'example.application:person.nationalId',
-    defaultMessage: 'Kennitala',
-    description: 'Some description',
-  },
-  age: {
-    id: 'example.application:person.age',
-    defaultMessage: 'Aldur',
-    description: 'Some description',
-  },
-  email: {
-    id: 'example.application:person.email',
-    defaultMessage: 'Netfang',
-    description: 'Some description',
-  },
-  phoneNumber: {
-    id: 'example.application:person.phoneNumber',
-    defaultMessage: 'Símanúmer',
-    description: 'Some description',
-  },
-  career: {
-    id: 'example.application:career',
-    defaultMessage: 'Starfsferill',
-    description: 'Some description',
-  },
-  history: {
-    id: 'example.application:history',
-    defaultMessage: 'Hvar hefur þú unnið áður?',
-    description: 'Some description',
-  },
-  careerHistory: {
-    id: 'example.application:careerHistory',
-    defaultMessage: 'Hefurðu unnið yfir höfuð einhvern tímann áður?',
-    description: 'Some description',
-  },
-  careerHistoryCompanies: {
-    id: 'example.application:careerHistoryCompanies',
-    defaultMessage: 'Hefurðu unnið fyrir eftirfarandi aðila?',
-    description: 'Some description',
-  },
-  future: {
-    id: 'example.application:future',
-    defaultMessage: 'Hvar langar þig að vinna?',
-    description: 'Some description',
-  },
-  dreamJob: {
-    id: 'example.application:dreamJob',
-    defaultMessage: 'Einhver draumavinnustaður?',
-    description: 'Some description',
-  },
-  yesOptionLabel: {
-    id: 'example.application:yes.option.label',
-    defaultMessage: 'Já',
-    description: 'Some description',
-  },
-  noOptionLabel: {
-    id: 'example.application:no.option.label',
-    defaultMessage: 'Nei',
-    description: 'Some description',
-  },
-  governmentOptionLabel: {
-    id: 'example.application:government.option.label',
-    defaultMessage: 'The government',
-    description: 'Some description',
-  },
-})
