--- libs/application/templates/children-domicile-transfer/src/lib/ReferenceApplicationTemplate.ts
@@ -1,159 +1,0 @@
-import {
-  ApplicationTemplate,
-  ApplicationTypes,
-  ApplicationContext,
-  ApplicationRole,
-  ApplicationStateSchema,
-  Application,
-} from '@island.is/application/core'
-import * as z from 'zod'
-
-const nationalIdRegex = /([0-9]){6}-?([0-9]){4}/
-
-type Events =
-  | { type: 'APPROVE' }
-  | { type: 'REJECT' }
-  | { type: 'SUBMIT' }
-  | { type: 'ABORT' }
-
-enum Roles {
-  APPLICANT = 'applicant',
-  ASSIGNEE = 'assignee',
-}
-const ExampleSchema = z.object({
-  person: z.object({
-    age: z.string().refine((x) => {
-      const asNumber = parseInt(x)
-      if (isNaN(asNumber)) {
-        return false
-      }
-      return asNumber > 15
-    }),
-    name: z.string().nonempty().max(256),
-    nationalId: z.string().refine((x) => (x ? nationalIdRegex.test(x) : false)),
-    phoneNumber: z.string().min(7),
-    email: z.string().email(),
-  }),
-  careerHistory: z.enum(['yes', 'no']).optional(),
-  careerHistoryCompanies: z
-    .array(
-      // TODO checkbox answers are [undefined, 'aranja', undefined] and we need to do something about it...
-      z.union([z.enum(['government', 'aranja', 'advania']), z.undefined()]),
-    )
-    .nonempty(),
-  dreamJob: z.string().optional(),
-})
-
-const ReferenceApplicationTemplate: ApplicationTemplate<
-  ApplicationContext,
-  ApplicationStateSchema<Events>,
-  Events
-> = {
-  type: ApplicationTypes.CHILDREN_DOMICILE_TRANSFER,
-  name: 'Reference application',
-  dataSchema: ExampleSchema,
-  stateMachineConfig: {
-    initial: 'draft',
-    states: {
-      draft: {
-        meta: {
-          name: 'Umsókn um ökunám',
-          progress: 0.33,
-          roles: [
-            {
-              id: Roles.APPLICANT,
-              formLoader: () =>
-                import('../forms/ChildrenDomicileTransferForm').then((module) =>
-                  Promise.resolve(module.ChildrenDomicileTransferForm),
-                ),
-              actions: [
-                { event: 'SUBMIT', name: 'Staðfesta', type: 'primary' },
-              ],
-              write: 'all',
-            },
-          ],
-        },
-        on: {
-          SUBMIT: {
-            target: 'inReview',
-          },
-        },
-      },
-      inReview: {
-        meta: {
-          name: 'In Review',
-          progress: 0.66,
-          roles: [
-            {
-              id: Roles.ASSIGNEE,
-              formLoader: () =>
-                import('../forms/ReviewApplication').then((val) =>
-                  Promise.resolve(val.ReviewApplication),
-                ),
-              actions: [
-                { event: 'APPROVE', name: 'Samþykkja', type: 'primary' },
-                { event: 'REJECT', name: 'Hafna', type: 'reject' },
-              ],
-              write: { answers: ['careerHistoryCompanies'] },
-              read: 'all',
-            },
-            {
-              id: Roles.APPLICANT,
-              formLoader: () =>
-                import('../forms/PendingReview').then((val) =>
-                  Promise.resolve(val.PendingReview),
-                ),
-              read: 'all',
-            },
-          ],
-        },
-        on: {
-          APPROVE: { target: 'approved' },
-          REJECT: { target: 'rejected' },
-        },
-      },
-      approved: {
-        meta: {
-          name: 'Approved',
-          progress: 1,
-          roles: [
-            {
-              id: Roles.APPLICANT,
-              formLoader: () =>
-                import('../forms/Approved').then((val) =>
-                  Promise.resolve(val.Approved),
-                ),
-              read: 'all',
-            },
-          ],
-        },
-        type: 'final' as const,
-      },
-      rejected: {
-        meta: {
-          name: 'Rejected',
-          roles: [
-            {
-              id: Roles.APPLICANT,
-              formLoader: () =>
-                import('../forms/Rejected').then((val) =>
-                  Promise.resolve(val.Rejected),
-                ),
-            },
-          ],
-        },
-      },
-    },
-  },
-  mapUserToRole(
-    id: string,
-    application: Application,
-  ): ApplicationRole | undefined {
-    if (application.state === 'inReview') {
-      return Roles.ASSIGNEE
-    }
-    return Roles.APPLICANT
-  },
-}
-
-export default ReferenceApplicationTemplate
